util.py split to autocrop.py, clus.py, param.py, vis.py; tested
