+++ test/consensus/ISS_rnd0_ch3_z0_blacked.csv
@@ -1,6 +1,6 @@
 centroid_x,centroid_y
-1.920000000000000000e+02,3.500000000000000000e+02
 1.797500000000000000e+02,3.505000000000000000e+02
+1.920000000000000000e+02,3.500000000000000000e+02
 1.470000000000000000e+02,8.780000000000000000e+02
 2.060000000000000000e+02,5.810000000000000000e+02
 1.700000000000000000e+02,3.590000000000000000e+02
