+++ test/test_crops/ISS_rnd0_ch3_z0_1.csv
@@ -1,5 +1,5 @@
-9.420000000000000000e+02
-1.389000000000000000e+03
-4.680000000000000000e+02
-1.036000000000000000e+03
-3.109619686800894733e+00
+2.310000000000000000e+02
+9.630000000000000000e+02
+1.940000000000000000e+02
+6.670000000000000000e+02
+1.898907103825136611e+00
