+++ test/test_crops/ISS_rnd0_ch3_z0_3.csv
@@ -1,5 +1,5 @@
-8.040000000000000000e+02
-1.388000000000000000e+03
-5.000000000000000000e+00
-4.580000000000000000e+02
-2.380136986301369717e+00
+2.000000000000000000e+02
+9.200000000000000000e+02
+6.720000000000000000e+02
+1.030000000000000000e+03
+1.930555555555555580e+00
